Fixed issue in SetHighPoint, implemented SetLowPoint
Fixed an issue in SetHighPoint.VI where the CR character was not being added to the end of the write command. This was causing a buffer overflow error in the transmit buffer.

Implemented SetLowPoint.VI (this VI is just a copy paste with an edit on the command structure of SetHighPoint with the commands changed).