Fixed issue on setting low point and high point
Fixed an issue with the location of the CR character in the write command for both the set low point and set high point VIs. Verified both VIs at the chamber (NOTE: read will occasionally return a invalid command string ("01C80532"), the value still is sent to the chamber in these instances.