Main Program Initial Phase Complete
Completed the initial phase of the main program. This is a recreation of the original software with a few extra features - asymmetric chamber dwell times, ability to choose which chamber to start in.

There are multiple fixes to the utility VIs used by the main that are not documented here - all utility VIs are documented.